Bug: rootDir wasn't the same value when --runInBand vs parallel
rootDir was modified after initialization to be process.cwd. In case of --runInBand,
The resolver was intialized before the rootDir value was altered. In case of parallel,
the Resolver was initialized in each spawned process using values of rootDir after
initialization.

Fix is to introduce a new config.cwd option that will be set to the parent processes
cwd. Adjust the console reporters to favor config.cwd over config.rootDir.